refactor: move CountDocument logic into collection API
