+++ test/integration/crud/abstract_operation.test.ts
@@ -54,11 +54,6 @@ describe('abstract operation', function () {
         correctCommandName: 'count'
       },
       {
-        subclassCreator: () => new mongodb.CountDocumentsOperation(collection, { a: 1 }, {}),
-        subclassType: mongodb.CountDocumentsOperation,
-        correctCommandName: 'aggregate'
-      },
-      {
         subclassCreator: () => new mongodb.CreateCollectionOperation(db, 'name'),
         subclassType: mongodb.CreateCollectionOperation,
         correctCommandName: 'create'
